create annotator URL
